Move receiving and buffering to separate module
